agt/pyproject.toml
[tool.poetry]
name = "agt"
version = "0.1.0"
description = "an agentic workflow maker"
authors = ["Ashutosh Upadhyay"]
license = "Private bitches"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.42.2"
langchain = "^0.3.19"
langchain-groq = "^0.2.4"
langchain-community = "^0.3.18"
langgraph = "^0.3.2"
replicate = "^1.0.4"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
